Allow to replace nonterminal symbols by numbers in the parser output.
If the new command line option '-r' is given, the generated parser
uses numeric codes instead of the full nonterminal symbols in the
parse tree.  This reduces the size of the parsing tables and also may
allow for faster parsing.  A new class attribute 'nonterminals'
contains a dictionary mapping numeric codes to the corresponding
symbols.